// src/types/enums/permissions.enum.ts
export enum Permissions {
  CAN_SUBMIT_REQUEST = 'can_submit_request',
  CAN_SUBMIT_WFH_REQUEST = 'can_submit_wfh_request',
  CAN_MANAGE_MY_REQUEST = 'can_manage_my_request',
  CAN_MANAGE_MY_WFH_REQUEST = 'can_manage_my_wfh_request',
  CAN_MANAGE_APAC_REQUESTS = 'can_manage_apac_requests',
  CAN_MANAGE_APAC_WFH_REQUEST = 'can_manage_apac_wfh_request',
  CAN_APPROVE_APAC_REQUEST = 'can_apac_approve_request',
  CAN_APPROVE_APAC_WFH_REQUEST = 'can_approve_apac_wfh_request',
  CAN_DISAPPROVE_APAC_REQUEST = 'can_apac_disapprove_request',
  CAN_DISAPPROVE_APAC_WFH_REQUEST = 'can_disapprove_apac_wfh_request',
  CAN_COMMENT_APAC_REQUEST = 'can_apac_comment_request',
  CAN_COMMENT_APAC_WFH_REQUEST = 'can_comment_apac_wfh_request',
  CAN_MANAGE_USERS = 'can_manage_users',
  CAN_ADD_USER = 'can_add_user',
  CAN_UPDATE_USER = 'can_update_user',
  CAN_DELETE_USER = 'can_delete_user',
  CAN_VIEW_LEAVE_BANK = 'can_view_leave_bank',
  CAN_MANAGE_LEAVE_BANK = 'can_manage_leave_bank',
  CAN_MANAGE_CANDIDATES = 'can_manage_candidates',
  CAN_VIEW_COMPANY_RESOURCE = 'can_view_company_resource',
  CAN_ADD_COMPANY_RESOURCE = 'can_add_company_resource',
  CAN_UPDATE_COMPANY_RESOURCE = 'can_update_company_resource',
  CAN_DELETE_COMPANY_RESOURCE = 'can_delete_company_resource',
  CAN_MANAGE_DEPARTMENTS = 'can_manage_departments',
  CAN_MANAGE_IT_REQUEST = 'can_manage_it_requests',
  CAN_MANAGE_LD_REQUEST = 'can_manage_ld_requests',
}